+++ include/ignition/rendering/Scene.hh
@@ -847,6 +847,10 @@ namespace ignition
       /// \return The created grid
       public: virtual GridPtr CreateGrid() = 0;
 
+      /// \brief Create new marker geometry.
+      /// \return The created marker
+      public: virtual MarkerPtr CreateMarker() = 0;
+
       /// \brief Create new text geometry.
       /// \return The created text
       public: virtual TextPtr CreateText() = 0;