+++ ogre2/src/Ogre2Scene.cc
@@ -335,6 +335,14 @@ GridPtr Ogre2Scene::CreateGridImpl(unsigned int /*_id*/,
 }
 
 //////////////////////////////////////////////////
+MarkerPtr Ogre2Scene::CreateMarkerImpl(unsigned int /*_id*/,
+    const std::string &/*_name*/)
+{
+  // TODO(anyone)
+  return MarkerPtr();
+}
+
+//////////////////////////////////////////////////
 TextPtr Ogre2Scene::CreateTextImpl(unsigned int /*_id*/,
     const std::string &/*_name*/)
 {